requirements-VA.txt
# DR-VA: Diabetic Retinopathy Visual Analytics Prototype
# Install: pip install -r requirements.txt

# Web framework
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0

# Data + ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
umap-learn>=0.5.4
Pillow>=10.0.0

# Explainability (for real LIME — optional for UI testing)
lime>=0.2.0
scikit-image>=0.21.0

# Model loading (for real LIME explanations via export.pkl)
# These must match your training environment versions
fastai==1.0.61
torch>=1.7.0
torchvision>=0.8.0
